requirements.txt
asgiref~=3.7.2
astrapy==1.5.0
#astrapy==0.7.7
#llama-index-vector-stores-astra-db==0.1.5
llama-index==0.10.57
beautifulsoup4~=4.12.2
bugsnag~=4.6.0
fastapi~=0.110.1
google-api-python-client~=2.109.0
google-cloud-aiplatform
httpx==0.27.2
httpx[http2]==0.27.2
openai
python-dotenv~=1.0.0
ragstack-ai==1.1.0
tiktoken
uvicorn~=0.25.0
chainlit~=1.1.402
llama-index-readers-web
langchain-community
llama-index-embeddings-langchain
llama-index-llms-openai
botocore==1.34.70
###
requests
urllib3
browserbase==0.3.0
firecrawl-py
playwright
MainContentExtractor
newspaper3k
scrapfly-sdk
html2text
spider-client
trafilatura
unstructured
selenium
chromedriver-autoinstaller
lxml_html_clean
youtube_transcript_api